+++ CHANGELOG.md
@@ -1,3 +1,22 @@
+## 0.9.1
+ - Fixed video repeating when loop turned off (thanks @mattvot)
+
+## 0.9.0
+ - Added `stop` method.
+ - Added `pause` method.
+ - Added `resume` method.
+ - Fixed seekbar not resetting after replaying video.
+
+## 0.8.6
+ - Added `seek` method.
+
+## 0.8.2
+ - Fixed error in more Android versions
+
+## 0.8.1
+ - Added `disableSeek`, `pauseOnPress`, `fullScreenOnLongPress` (thanks @mateofd).
+ - Bumped react-native-video version (thanks @ndcollins).
+
 ## 0.8.0
  - Updated to react native 0.46.4 and fixed PropTypes (thanks @Traviskn).
  - Added `onStart` property (thanks @gvillenave).